Fix Duplicate Entries in Free Food Ad Campaign
When opening and closing the dropdown box multiple times while attempting to create an ad campaign for free food vouchers, the game would duplicate the available food items in the list each time.

I discovered that it was because the vector wasn't being cleared each time the list was populated it, so an easy fix is to clear it before continuing.