Explicitly make calls to hide disabled dynamic binding
Dynamic binding is disabled in constructors and destructors in C++, meaning that when virtual functions are called that could be overriden by derived classes, the member functions of those classes won't be called there.
This commit makes it more obvious that this is happening by explicitly calling the member function of the same class.