Guests "Heading to" ride skip checks like vanilla
Guests "Heading to" a ride will now skip some of the checks they'd otherwise perform that take into account data calculated about the ride, i.e. how much is sheltered during the rain, how intense it is, and how nauseating it is. In effect, they will not bounce off the queue entrance with thoughts "not going on RIDE while it's raining", "RIDE looks too intense for me", "I want to go on something more thrilling than RIDE", "Just looking at RIDE makes me sick". They are still likely to bounce off if the ride stats haven't been calculated and it's a ride capable of g-forces, or if it's value is uncalculate or too low vis-a-vis the ticket price.

A pre-existing comment in the code suggests that this is the intended behavior, and it's confirmed to be the behavior in at least RCT1 and RCT Classic.