Add Inline Twist to B&M Track
This adds the inline twist to the B&M track.

Feature is complete. Note: The inline twist doesn't draw correctly when going out of the tunnel. This is sadly an original bug (#7466) ( as this is also present on the Flying RC's Flying Twist. If you guys can still fix those that's awesome, but it's not entirely necessary!)

Merged Feature #16132 and #16389 on the Changelog