Fix SHGetPathFromIDListLongPath (#16545)
The function did not truncate the string afterwards, so you ended up with a very long string containing null terminators on the end. This wouldn't have caused an issue when passing it as a char *, but due to it now being passed as string or string_view, this causes issues.

It also wouldn't have worked for paths greater than MAX_PATH, because it wasn't passing the buffer size to the function.