Adds hidden players & makes headless servers hidden players (#5141)
Hidden players are normally not visible in the player list and do not add to
the player count (nor do they count towards the maximum amount of players).
Although the code supports having any player as a hidden player, there is
currently no way for any player other than a headless server to become hidden.

Hidden players can be seen in the player list if the "debugging tools" option
is enabled.

The main effect of this commit is that empty headless servers will have a
player count of 0 instead of 1.