Fix implicit floating point cast imprecision error
"error: implicit conversion from 'unsigned int' to 'float' changes value from 4294967295 to 4294967296"
This value was picked since it's the maximum value that the mersenne twister engine returns. It happens to be too large for floating points to store the value accurately. The previous number is 256 lower, and the next number is be 512 higher. This is what the compiler warned about.
Strangely, this error only showed up on a later PR after this code was introduced.

This fix essentially limits the returned random values to 2 bytes, which floats can cover in its entirety, and should not have any noticeable effect on its distribution.