Remove unused preprocessor placeholder
This define used to be `__DATE__ " " __TIME__`, but was replaced by a placeholder in ad687a7ecf185fd7df5932f1e2ae2e65ef0360a5. As mentioned there, it was intended to be set by the build tool, but this was never done. Meanwhile, we already have the git commit hash and branch information logged right above this, so this define doesn't really serve any purpose anymore.