Implement "Same price by default" option
This is a quality of life feature intended to compliment the "Auto open
shops" setting. It automatically checks "Same price throughout park" for
newly created shops and stalls. For predictability/compatibility, the behavior
needs to meet these conditions:

1. Ride object must be either a shop or restroom. Excludes ride photos.
2. Shop item's "Same Price" option shouldn't be checked
3. The shop/stall shouldn't already exist in the active park map. This
   prevents accidentally overriding shop prices of saved games and new scenarios.
4. The item itself should also not exist in the park. For example,
   when placing Souvenir Shop 1 with an Info Kiosk present, the Cuddly
   Toy will have "Same Price" checked while the Umbrella will remain
   unchecked.

In practice, this setting combined with "Auto Open Shops" enables rapid
placement of shops & stalls and reduces the overhead of micromanaging
after the fact.

Most of this logic occurs inside the `RideCreateAction` and `ShopItem`.
This being my first time in this codebase (and C++ lol) there may be a
better implementation, like setting all `SamePriceThroughoutPark` flags
at once on scenario load.