Remove Reverse Incline Launched Shuttle mode from Junior RC
This mode was not available in vanilla RCT2. We added it for RCT1 parity
before realising it was a better idea to just spin it off into its own
roller coaster (namely the Classic Mini RC). When we did eventually split
it, I overlooked this one.

Obviously, this mode still remains available on the Classic Mini RC.