Fix #17889: Peeps don't stop at railway crossings
If peeps approached a railway crossing from downhill, they would not stop for a passing train. This was because the destination z wasn't up-to-date.

Moving the logic to the generic peep_footpath_move_forward method, where the new target destination is determined, resolves this issue. It also prevents some code duplication in the 'Guest' and 'Peep' classes.