Fix #18063: Guests stopping at railway crossings
Guests could stop at railway crossings for actions like taking a picture, watching a ride / a ride being constructed, reading a map etc. This also could mean they'd be in the way for a passing train.

Now, these actions will no longer be performed if the guest is at a crossing.