Update the Bug Report Form
This is to accomodate for new types of issues that could have been introduced with the New Save Format, Asset Packs, the ability to load in RCT Classic and lastly if the issue is related to zfighting.

I don't know if it's smart to automate more labelling or keep that to the devs, but I don't know how to do so if needed.