Improve OpenGL Rectangle SetInstances Performance
Change to use glBufferSubData instead of glBufferData

An initial buffer size is set with glBufferData.
glBufferSubData is then used to copy to the existing buffer,
rather than creating a new one for each frame.

We keep track of the buffer size, and if instances size increases
above the buffer size, a new buffer is created with glBufferData