Fix #18467: Update FilterSelected logic when the active screen is TM.
The Track Manager has no concept of selectable items. Therefore we shouldn't be concerned if the SELECTED or NONSELECTED filters are currently active in the global filters list when on this screen and, as such, those filters should be ignored when determining which items to display in the item list.