Fix #17069: Fix build on ppc64el architectures
For a user defined floating point literal, the parameter type must be a
`long double` which is problematic on ppc64el, as the architecture uses a
pair of `doubles` to represent that type. This cannot be converted to a
`constexpr`. As a workaround, statically cast the `long double` down to a
`double`. All of the uses of _GBP constants fit just fine, and if anyone
really tries to use a gigantic constant that can't fit in a double, they are
probably going to be breaking other things anyways.
For more details, see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=26374

The Debian packaging of openrct2 has included this patch since 0.4.0, with
no reported issues.

Signed-off-by: Mathias Gibbens <mathias@calenhad.com>