Fix #19553: Undefined upkeep cost check uses new sentinel value
The datatype for ride.upkeep_cost was changed from money16 to money64 in
PR #19469, and so did the sentinel value used to represent an undefined
money value, but there was a check that was not updated to use this new
sentinel value. As a result, rides with undefined upkeep costs end up
contributing to the monthly ride running costs, causing the following...

- The overall cash balance to be wrong
- The cash value is shown red (as the internal value is indeed negative)
  but the displayed value is shown positive (possibly due to truncation)
- The graphs shows a large dip, but the y-axis is no longer monotonic
  (possibly due to truncation when displaying the y-axis values).

The primary fix is to update to the new sentinel value.
Not fixed - handling of extreme values when displaying/formatting
currency.

Testing done...

- Confirmed that the symptoms are reproducible before this change.
- Confirmed that the symptoms no longer occured after this change using
  the same save file throughout one month cycle.