CMakeLists.txt: Do not force download of OpenMSX/OpenSFX
This change removes the hardcoded enable for Linuxsystems, where
DOWNLOAD_OPENMSX and DOWNLOAD_OPENSFX is always set to ON and as a
result overwriting users choice not to download those files.

Also setting DOWNLOAD_OPENMSX and DOWNLOAD_OPENSFX to ON, in order not
to break existing behauvior.

This fixes https://github.com/OpenRCT2/OpenRCT2/issues/19763
Signed-off-by: Conrad Kostecki <conikost@gentoo.org>