Return destination tolerance for entering
Reset the destination tolerance back to 7 for peeps entering
the park.  This had some unforseen side-effects on the rest of
the pathfinding.  Increasing the z by 2 specifically for the
peeps walking through a park entrance to an upward slope
also solves the issue of them not making it to the path,
so I will keep this as a better solution to the issue for now.