Restores original intent of #22199
nulls are now considered higher values than anything, making them appear on the tail of ascending lists again

checking for NULL in ratings.intensity and ratings.nausea didn't work, they seem to retain there values pre ratings reset and setting them to null would change the behaviour of the game - so instead we check for ratings.isNull() in the sorting to properly put null ratings to tail of sorted lists.

Tested in-game.