Quick patch for About screen (#23117)
It seems there's a bigger issue going on atm. This fixes the problem for now, but when we add more names we will have to extend this more. Reason being that the it centers the text in the line, rather than extending it further downwards.
I personally do not know how to fix that, but this should suffice for now.