Fix #23797: 3D sign text from older parks cut off
The origin of this bug is an issue in the logic for determining
when to wrap 3D sign text on multi-line signs in Paint.LargeScenery.cpp.
When processing the second line, no check was made that the full text
could fit within the sign's bounding box, and so spaces or newlines
were always used to wrap to a (nonexistent) third line, even when
the text would otherwise fit.

This fix addresses the issue by adding a check for this case, when
we're at the end of the text string and the max text width still
hasn't been exceeded. In that case, the best display string is set
to the full current string, rather than the string up to the last
space glyph.