Fix #18433: TTF fonts can not be located when font names are translated
The issue here is that fontconfig may report the name of a given
font face in a language specified by the current locale, while the
OpenRCT2 font code checks the result against an English name.

This fix addresses the issue by temporarily changing the environment
LANG variable to use the default C locale, which should guarantee
English/ASCII names are returned by fontconfig. It uses an RAII
mechanism to ensure that the modified environment variable is restored
to its original value on exiting GetFontPath.