Fix #23868: Game crashes when moving to a corner of my park
Based on the discussion of this issue, it appears to be caused by
an oversight in the 3D sign text cutoff fix (7de61c7) where in
certain cases the index can wrap around and cause invalid input
to substr. The documentation on what happened in this case was
ambiguous, but in this case it leads to an exception. The fix is
to have a check for a zero index before decrementing in the line
wrapping code.