24146 - Add bloons, money and sounds to plugin
Add the colouring of balloons, as well as getting all balloons in
the park, or on a tile for plugins.
Add setting a value to a money effect, as well as getting all
the money effects in the park, or on a tile for plugins.
Add a function to play sounds from a plugin.