Build G2.dat when building the Windows exe under VS
By moving the target into the .vcxproj and adjusting some values we will run the g2 step as part of every build of the vcxproj. This will continue to always run when the msbuild project is run (although it will run before the tests). In the VS IDE where this never ran it will now run when the exe is built/rebuilt. The g2 step will not always run in the VS IDE because the IDE's up-to-date check will not consider the input or output as part of the build. Fixing that requires further work either by using tlog files or breaking this all out to a new proj file in the sln.