Add scripting hooks for ride breakdowns and fixes
This adds the hook "ride.breakdown" to the scripting api.
It is called whenever a ride breaks down. The hook includes
the id of the ride that broke down, the type of breakdown,
and what car broke down, if applicable.

This also adds the hook "ride.fix" to the scripting api.
It is called whenever a ride is fixed. The hook includes
the id of the ride that was fixed.

I left some TODO comments in the code for things I wasn't
sure about, particularly when the "ride.breakdown" hook
is invoked. It is currently invoked within the "RidePrepareBreakdown"
function in openrct2/ride/Ride.cpp, which is called before the breakdown
notification is shown to the player. I could change it so that the hook
is fired at the same time that the notification is sent if that
is preferred. Note that the notification does not fire immediately when
forcing a breakdown via the drop-down menu in the maintenance tab for
a ride.

Lastly, I changed my username under contributors.md to reflect my new
account.

An example plugin that uses the new hooks may be found here:
https://pastebin.com/GmhYwGsb