Linux: add .desktop hint to request discrete GPU (#24563)
This hint was added in v1.4 of the spec and supported by GNOME (since 2020) and KDE (since 2021) and is intended for systems with hybrid graphics. It tells the system to launch the application on the more powerful discrete GPU as opposed to the more energy efficient integrated graphics.

Systems like this (e.g. the Optimus laptops) are the spawn of Satan, but this makes the experience much more enjoyable for the poor users with a system like this.

Reference: https://specifications.freedesktop.org/desktop-entry-spec/latest/recognized-keys.html