Fix #24904: gracefully handle incomplete CSS files (#24923)
* Fix #24904: gracefully handle incomplete CSS files

Some CSS files are truncated, possibly due to users overwriting RCT2
assets with parts of RCT1. This causes incorrect access to freed memory
and potential double frees.

This change verifies CSS data has expected samples and if not, informs
user about incorrect assets. The samples that exist are still loaded and
available for use.

There was no backtrace report generated for this on Windows.

* Skip reading of unused variables