Change land tool to account for sloped track and paths
This allows you to modify the land underneath sloped track and paths without the zero clearances cheat.

Closes #2296
Closes #2307

https://github.com/user-attachments/assets/8cb51199-a5b4-48e4-9cc3-34fd6ce7410e

It passes the slope in to the function that checks for construction clearances, and calculates whether it fits under the track or path.

In some very rare cases it is possible to build some track pieces in an arguably invalid way over 2 height sloped land, this does not let you raise the land in to that invalid position after the track piece has been built.