Fix reverse freefall photo using the wrong track sprite
This fixes the on ride photo of the reverse freefall coaster having incorrectly coloured ties due to using a sprite from the air powered coaster. This uses the station sprite from the reverse freefall coaster instead.

This removes this special case in the paint code. If I remember right, it's the only function that changes colors in this way.