Restore maximum ride price algo to vanilla
This reverses the functional changes made in #5294. While the new
algorithm works out well for roller coaster and encourages better
rides in general, it is hampered by the fact that the value of
flat rides is abysmal and this means that you can never ask a
decent price.

This should probably be fixed at the source instead of
papered over by an absolute bonus, but fixing that is a lot more
delicate and complicated than restoring vanilla’s behaviour.
Something to revisit at a later date and with sufficient testing.