Allow shifting track designs underground
Since about a year ago, OpenRCT2 has allowed players to shift ride/track designs using modifier keys (#22669). This was, however, limited to heights above-ground. This PR changes this by omitting the 'Z placement assist' when ctrl/shift modifiers are used. This means the helpful 'nudge' to overground is now omitted if (and only if) these modifier keys are involved.

Below is an example of what placement now looks like (with clearance checks disabled for easier debugging):

https://github.com/user-attachments/assets/3b23bc1d-ab7b-43a2-ad40-216ab83952d2