Fix multithreaded access to additional crash files (#25271)
`GetScenarioInfo` can be called from a `JobPool` in `FileIndex::Build`,
resulting in multiple threads trying to access the `_uploadFiles` map
for breakpad.

This commit adds a mutex guard for `_uploadFiles` and provides unique
names for the entries.

The crash handler itself does not verify mutex state, possibly causing
data races, but I don't think it's safe to access it at this point and
we have to make do with whatever state is present at this point.